requirements.txt
# Runtime dependencies for ekigega_backend
# Adjust exact versions for your environment; these are conservative minimums.

Django>=4.2,<6.0
djangorestframework>=3.14
drf-yasg>=1.21.0
dj-database-url>=3.0.1
django-cors-headers>=3.14.0
django-filter>=23.1
djangorestframework-simplejwt>=5.2.0
django_celery_results>=2.6.0
django-redis>=5.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
celery==5.6.2
redis==7.1.0
pillow>=10.0.0
waitress>=2.1.0
gunicorn>=20.1.0
whitenoise>=6.11.0
scikit-learn>=1.3.0
pandas>=2.0.0
python-dateutil>=2.8.0

# Production servers:
# - Windows/dev: use Django runserver (python manage.py runserver) or waitress
# - Linux/production: use gunicorn (pip install gunicorn)
# If you use other DB backends, add the appropriate driver (e.g., mysqlclient)
